Working on u256 profiling
