+++ zisk-programs/revm/Cargo.toml
@@ -20,5 +20,11 @@ ruint.workspace = true
 
 guest-reth.workspace = true
 
+[features]
+# Force the u256 wrappers to use the ruint fallback even when targeting zisk
+ruint-fallback = []
+# Wrap every u256 function with ziskos profiling functions
+profile = []
+
 [lints]
 workspace = true
